Solution to part 1
I really messed up on my understanding of pointers for methods.
Finally sorted it out and got to the answer, although it's not
very fast (~19 seconds). Brute force obvs isn't the way to go
and with some more analysis I could probably optimize. We'll see
if part2 forces me to do that.